+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This is the changelog file of My Games (games.anikethchavare.com).
 
 **Note:** All the dates below are in the IST time format.
 
+### <i>Commit 3 (7th April 2025)</i>
+
+1. Set a common favicon for all web pages under "media/favicon.png".
+2. Added an "Error 404" page for unknown requests.
+3. Added a "robots" meta tag for all web pages.
+4. Updated the response headers for all web pages by adding the following: X-Frame-Options, Cross-Origin-Opener-Policy, Permissions-Policy, Cache-Control, X-Content-Type-Options, Referrer-Policy, and Strict-Transport-Security.
+5. Updated the response header "Content-Security-Policy" for better security.
+
 ### <i>Commit 2 (6th April 2025)</i>
 
 1. Added a new game "Rock Paper Scissors", in short as "RPS", to the route "/rps".